pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = ""
authors = ["Ahmed Shahwan <ahmed@shahwan.me>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.1.2"
psycopg2-binary = "^2.8.6"
django-environ = "^0.4.5"
sentry-sdk = "^0.19.1"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
django-model-utils = "^4.0.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.0.0"}
djangorestframework = "^3.12.1"
django-extensions = "^3.0.9"
factory-boy = "^3.1.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
flake8-black = "^0.2.1"
flake9 = "^3.8.3"
docker-compose = "^1.27.4"
ipdb = "^0.13.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs        # exclude a few common directories in the
    | \.git         # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | venv,
    | \.env,
    | env,
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | __pycache__
  )/
  | foo.py          # also separately exclude a file named foo.py in
                    # the root of the project
)
'''

[tool.flake8]
max-line-length = 88
max-complexity = 25
select = 'C,E,F,W,B,B950'
ignore = 'E133,E501,W503,E203,W605'
hang-closing = true
exclude = '''
    .git,
    migrations,
    __pycache__,
    manage.py,
    templates,
    dist,
    build,
    venv,
    .venv,
    .env,
    env,
    node_modules,
    .svn,
    CVS,
    .bzr,
    .hg,
    .git,
    __pycache__,
    .tox,
    .eggs,
    *.egg
'''
